+++ services/QuillGrammar/built/index.js
@@ -1,5 +1,5 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import App from "./components/App";
+import App from "./App";
 ReactDOM.render(React.createElement(App, null), document.getElementById("app"));
 //# sourceMappingURL=index.js.map
