+++ services/QuillGrammar/package.json
@@ -27,6 +27,7 @@
     "@types/react-router-config": "^1.0.6",
     "@types/react-router-dom": "^4.2.5",
     "antd": "^3.3.1",
+    "node-sass": "^4.9.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-redux": "^5.0.7",
