--- services/QuillGrammar/src/components/layouts/Header.less
@@ -1,10 +1,0 @@
-.user-logout {
-    float: right;
-    width: 100px;
-    margin-top: 8px;
-    margin-right: 32px;
-}
-
-.ant-layout-header {
-  height: 58px!important;
-}
