--- services/QuillGrammar/src/components/layouts/PageLayout.less
@@ -1,8 +1,0 @@
-#app,
-div[data-reactroot] {
-  height: 100%;
-}
-.ant-layout {
-  background-color: #F3F9F9;
-  height: 100%;
-}
