--- services/QuillGrammar/src/components/layouts/Sidebar.less
@@ -1,35 +1,0 @@
-.ant-layout-logo {
-  height: 32px;
-  background: #333;
-  border-radius: 6px;
-  margin: 16px;
-  text-align: center;
-}
-.sider-trigger {
-  position: absolute;
-  top: 15px;
-  right: -25px;
-  width: 25px;
-  height: 25px;
-  line-height: 25px;
-  text-align: center;
-  color: #fff;
-  background: #404040;
-  border-radius: 0 4px 4px 0;
-  cursor: pointer;
-  .anticon {
-    line-height: 25px;
-  }
-}
-.ant-layout-sider-collapsed .ant-menu .anticon {
-  font-size: 16px;
-  margin-left: 8px;
-}
-
-.ant-layout-sider-collapsed .nav-text {
-  display: none;
-}
-
-.ant-layout-sider-collapsed .ant-menu-submenu-vertical > .ant-menu-submenu-title:after {
-  display: none;
-}
