+++ services/QuillGrammar/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import App from "./components/App";
+import App from "./App";
 
 ReactDOM.render(
     <App />,
