+++ services/QuillGrammar/src/layouts/Header.scss
@@ -1,0 +1,6 @@
+a {
+  color: white;
+}
+img {
+  height: 25px;
+}
