+++ services/QuillGrammar/src/layouts/Header.tsx
@@ -1,0 +1,19 @@
+import * as React from "react";
+import { Layout, Row } from "antd";
+import { Link } from "react-router-dom";
+import './Header.scss'
+
+export const Header: React.StatelessComponent<{}> = () => {
+    return (
+        <Layout.Header style={{
+          height: '60px',
+          width: "100%",
+          backgroundColor: "#00c2a2",
+          padding: "0 30px"}}>
+          <Row type="flex" align="middle" justify="space-between" style={{height: '100%'}}>
+            <img src="https://d2t498vi8pate3.cloudfront.net/assets/home-header-logo-8d37f4195730352f0055d39f7e88df602e2d67bdab1000ac5886c5a492400c9d.png" />
+            <a href="http://quill.org/profile">Save & Exit</a>
+          </Row>
+        </Layout.Header>
+    );
+};
