+++ services/QuillGrammar/src/layouts/PageLayout.tsx
@@ -1,16 +1,13 @@
 import * as React from "react";
 import { Layout } from "antd";
-import Sidebar from "./Sidebar";
 import { Header } from "./Header";
 import { Redirect } from "react-router-dom";
 import {renderRoutes} from "react-router-config";
 import { routes } from "../routes";
-import "./PageLayout.less";
 
 const PageLayout: React.StatelessComponent<{}> = () => {
     return (
-        <Layout className="ant-layout-has-sider">
-            <Sidebar />
+        <Layout className="ant-layout">
             <Layout>
                 <Layout.Content>
                     <Header />
