+++ services/QuillGrammar/src/routes.tsx
@@ -1,12 +1,17 @@
 import * as React from "react";
 import { RouteConfig } from "react-router-config";
 import { Route, Switch } from "react-router-dom";
-import HomePage from "./pages/HomePage";
-// import { TodoPage } from "./pages/TodoPage";
-import AboutPage from "./pages/AboutPage";
+import HomePage from "./components/HomePage";
+// import { TodoPage } from "./components/TodoPage";
+import AboutPage from "./components/AboutPage";
 import PageLayout from "./layouts/PageLayout";
+import PlayGrammar from "./components/PlayGrammar"
 
 export const routes: RouteConfig[] = [
+    {
+      path: "/play/sw",
+      component: () => (<PlayGrammar />)
+    },
     {
         path: "/home",
         exact: true,
